Adicionado controllers de Aluno, Coordenador, Disciplina e Equipamento.

+++ Backend/src/main/java/com/br/sga/domain/Coordenador.java
@@ -1,13 +1,7 @@
 package com.br.sga.domain;
 
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.SequenceGenerator;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -27,7 +21,7 @@ public class Coordenador implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_coordenador")
     @SequenceGenerator(name = "seq_coordenador", sequenceName = "seq_coordenador", allocationSize = 1)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "nome")
     private String nome;
