+++ Backend/src/main/java/com/br/sga/domain/Periodo.java
@@ -1,15 +1,13 @@
 package com.br.sga.domain;
 
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.SequenceGenerator;
 import jakarta.persistence.Table;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Column;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -30,7 +28,7 @@ public class Periodo implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_periodo")
     @SequenceGenerator(name = "seq_periodo", sequenceName = "seq_periodo", allocationSize = 1)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "descricao", nullable = false)
     private String descricao;
