+++ Backend/src/main/java/com/br/sga/repository/AuditoriaEventoRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.AuditoriaEvento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditoriaEventoRepository extends JpaRepository<AuditoriaEvento, Integer> {
+public interface AuditoriaEventoRepository extends JpaRepository<AuditoriaEvento, Long> {
 }
