+++ Backend/src/main/java/com/br/sga/repository/CoordenadorRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Coordenador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoordenadorRepository extends JpaRepository<Coordenador, Integer> {
+public interface CoordenadorRepository extends JpaRepository<Coordenador, Long> {
 }
