+++ Backend/src/main/java/com/br/sga/repository/CoordenadoriaRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Coordenadoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CoordenadoriaRepository extends JpaRepository<Coordenadoria, Integer> {
+public interface CoordenadoriaRepository extends JpaRepository<Coordenadoria, Long> {
 }
