+++ Backend/src/main/java/com/br/sga/repository/DisciplinaRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Disciplina;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisciplinaRepository extends JpaRepository<Disciplina, Integer> {
+public interface DisciplinaRepository extends JpaRepository<Disciplina, Long> {
 }
