+++ Backend/src/main/java/com/br/sga/repository/EquipamentoRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Equipamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquipamentoRepository extends JpaRepository<Equipamento, Integer> {
+public interface EquipamentoRepository extends JpaRepository<Equipamento, Long> {
 }
