+++ Backend/src/main/java/com/br/sga/repository/LocalRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Local;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalRepository extends JpaRepository<Local, Integer> {
+public interface LocalRepository extends JpaRepository<Local, Long> {
 }
