+++ Backend/src/main/java/com/br/sga/repository/PeriodoRepository.java
@@ -3,5 +3,5 @@ package com.br.sga.repository;
 import com.br.sga.domain.Periodo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeriodoRepository extends JpaRepository<Periodo, Integer> {
+public interface PeriodoRepository extends JpaRepository<Periodo, Long> {
 }
